Trait linera_wit_bindgen_host_wasmtime_rust::AllBytesValid

pub unsafe trait AllBytesValid { }

Unsafe trait representing types where every byte pattern is valid for their representation.

This is the set of types which wasmtime can have a raw pointer to for values which reside in wasm linear memory.